feat: setup localization module and move Nstack related stuff there
