+++ localization/src/main/res/values/nstack_keys.xml
