+++ settings.gradle.kts
@@ -18,3 +18,4 @@ rootProject.name = "android-template"
 include(":app")
 include(":core")
 include(":designsystem")
+include(":localization")
